+++ plume/src/main/kotlin/io/github/plume/oss/drivers/OverflowDbDriver.kt
@@ -22,10 +22,6 @@ import io.github.plume.oss.domain.mappers.VertexMapper.checkSchemaConstraints
 import io.github.plume.oss.domain.model.DeltaGraph
 import io.github.plume.oss.metrics.ExtractorTimeKey
 import io.github.plume.oss.metrics.PlumeTimer
-import io.github.plume.oss.util.ExtractorConst.TYPE_REFERENCED_EDGES
-import io.github.plume.oss.util.ExtractorConst.TYPE_REFERENCED_NODES
-import io.shiftleft.codepropertygraph.generated.NodeKeyNames.FULL_NAME
-import io.shiftleft.codepropertygraph.generated.NodeTypes.METHOD
 import io.shiftleft.codepropertygraph.generated.nodes.*
 import org.apache.logging.log4j.LogManager
 import overflowdb.*
@@ -198,8 +194,6 @@ class OverflowDbDriver internal constructor() : IDriver {
         return deepCopyGraph(Traversals.getMethodStub(graph, fullName))
     }
 
-    override fun getMethodNames(): List<String> = getPropertyFromVertices(FULL_NAME, METHOD)
-
     private fun deepCopyGraph(edges: List<Edge>): Graph {
         val newGraph = newOverflowGraph()
         PlumeTimer.measure(ExtractorTimeKey.DATABASE_READ) {
@@ -241,38 +235,6 @@ class OverflowDbDriver internal constructor() : IDriver {
                 }
         }
         return g
-    }
-
-    override fun getProgramTypeData(): Graph {
-        val pg = newOverflowGraph()
-        PlumeTimer.measure(ExtractorTimeKey.DATABASE_READ) {
-            graph.nodes(*TYPE_REFERENCED_NODES).asSequence().forEach { n ->
-                // Add vertices
-                if (pg.node(n.id()) == null) {
-                    val node = pg.addNode(n.id(), n.label())
-                    n.propertyMap().forEach { (key, value) -> node.setProperty(key, value) }
-                }
-            }
-            TYPE_REFERENCED_EDGES.forEach { tre ->
-                graph.edges(tre).asSequence()
-                    .map { e ->
-                        val srcNode = if (pg.node(e.outNode().id()) == null) {
-                            val s = pg.addNode(e.outNode().id(), e.outNode().label())
-                            e.outNode().propertyMap().forEach { (key, value) -> s.setProperty(key as String, value) }; s
-                        } else {
-                            pg.node(e.outNode().id())
-                        }
-                        val dstNode = if (pg.node(e.inNode().id()) == null) {
-                            val d = pg.addNode(e.inNode().id(), e.inNode().label())
-                            e.inNode().propertyMap().forEach { (key, value) -> d.setProperty(key as String, value) }; d
-                        } else {
-                            pg.node(e.inNode().id())
-                        }
-                        Triple(srcNode, dstNode, e.label())
-                    }.forEach { (src, dst, e) -> pg.node(src.id()).addEdge(e, pg.node(dst.id())) }
-            }
-        }
-        return pg
     }
 
     override fun getNeighbours(v: NewNodeBuilder): Graph = deepCopyGraph(Traversals.getNeighbours(graph, v.id()))